refactor: Homeのロジック部分を切り出し
